+++ .github/workflows/CHANGELOG.md
@@ -1,9 +1,10 @@
 # k8f 
 ## Changes
-- fixed profile and assume role
-- added dynamic support for STS aka assume role/ switch role
+- add support to STS/Switch role in both creds and config
+- user can set creds in both config and creds dynamicly
 ## Bugfix
-- fixed typo in Command options
+- [[support] add support for STS using only config file (no Credential file) #4
+](https://github.com/AdamRussak/k8f/issues/4)
 ## Known Issue
 - connect command with "dry-run" flag will default to Json insted of Yaml
 <!-- ## Bugfix -->